pyproject.toml
[tool.poetry]
name = "paper-search"
version = "0.1.0"
description = ""
authors = ["nomnomnonono <nozomu.koujiguchi1204@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
notebook = "^6.5.3"
requests = "^2.29.0"
bs4 = "^0.0.1"
pandas = "^2.0.1"
omegaconf = "^2.3.0"
transformers = "^4.28.1"
pydub = "^0.25.1"
numpy = "^1.24.3"
PySoundFile = "^0.9.0"
torch = "^2.0.0"
torchvision = "^0.15.1"
torchaudio = "^2.0.1"
tqdm = "^4.65.0"
fugashi = {extras = ["unidic-lite"], version = "^1.2.1"}
ipadic = "^1.0.0"
gradio = "^3.28.1"

[tool.poetry.dev-dependencies]
pysen = {version = "^0.10.3", extras = ["lint"]}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pysen]
version = "0.10.3"

[tool.pysen.lint]
enable_black = true
enable_flake8 = true
enable_isort = true
enable_mypy = false
mypy_preset = "strict"
line_length = 88
py_version = "py39"

[[tool.pysen.lint.mypy_targets]]
paths = ["."]
